Update patch set 7
Patch Set 7:

Running example_iris.R should create a model bundle file; one basic way to test it would be to simple test for the existence of the file after running it. Slightly more sophisticated would be to expand the file (it's a zip archive) and inspect the contents. Yet more sophisticated would be to try to run the model and supply inputs/check outputs. We could at least to the simplest thing.

Then there is the matter of pushing the code into a standard repo. We aren't creating a nexus artifact here, though, since the R client library isn't something that is a component of a running Acumos system per se; instead it is an R package that a modeler needs to install locally when onboarding R models.

This might need to be a manual step if we want to install into CRAN (the most standard repository of R packages), since I don't know a way to automate deployment there. Other places on the web that are world-accessible would be possible; we would just have to document the site/url from which the package could be onboarded/installed. I believe the current instructions point you to Rforge (http://rforge.net/acumos/); I'm not sure what credentials one would need to update that (it seems to be owned by Simon Urbanek), or if automated package update is possible there.

Patch-set: 7